Python interface might compile in cleanup
Former-commit-id: c6854b3f7bb3b22303bfdc7d53821cf17dbe32b4 [formerly 2ae1f82d8a2c7c6e7b7069f2b94e8be0061822a8]
Former-commit-id: a4601fe698ddfaf862eff318b8f110efc79b1771

+++ src/librapid/internal/forward.hpp
@@ -62,12 +62,12 @@ namespace librapid {
 	public:
 		StrOpt() : digits(-1), base(10), scientific(false) {}
 
-		StrOpt(i32 digits_, i32 base_, bool scientific_) :
+		StrOpt(i32 digits_, i8 base_, bool scientific_) :
 				digits(digits_), base(base_), scientific(scientific_) {}
 
 	public:
 		i32 digits;
-		i32 base;
+		i8 base;
 		bool scientific;
 	};
